# src/adhara/interop-service/crosschainInterop/api/admin/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Administration API",
    "description": "Crosschain interop service Administration API defines endpoints for maintaining the Corda Notaries, Corda Participants, Corda Registered Functions, Interop Authentication Parameters, Interop Participants, and Validators resources."
  },
  "host": "localhost:3031",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/{networkId}/validatorSetInstructions": {
      "get": {
        "tags": [
          "ValidatorSetInstructions"
        ],
        "summary": "Fetch the validator set instruction.",
        "description": "",
        "operationId": "getValidatorSetInstruction",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "operationId",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched the validator set instruction.",
            "schema": {
              "$ref": "#/definitions/ValidatorSetInstructionResponse"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "post": {
        "tags": [
          "ValidatorSetInstructions"
        ],
        "summary": "Create validator set instruction.",
        "description": "",
        "operationId": "postValidatorSetInstruction",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "validatorSetInstruction",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ValidatorSetInstructionRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created validator update instruction.",
            "schema": {
              "$ref": "#/definitions/ValidatorSetInstructionResponse"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      },
      "delete": {
        "tags": [
          "ValidatorSetInstructions"
        ],
        "summary": "Remove validator set instruction.",
        "description": "",
        "operationId": "deleteValidatorSetInstruction",
        "parameters": [
          {
            "name": "networkId",
            "in": "path",
            "required": true,
            "type": "number"
          },
          {
            "name": "operationId",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Removed validator set instruction.",
            "schema": {
              "$ref": "#/definitions/ValidatorSetInstructionResponse"
            }
          },
          "500": {
            "description": "An error occurred."
          }
        }
      }
    }
  },
  "definitions": {
...
    "EthereumProof": {
      "type": "object",
      "properties": {
        "networkId": {
          "type": "string",
          "example": "string"
        },
        "sourceNetworkId": {
          "type": "number",
          "example": 0
        },
        "encodedInfo": {
          "type": "number",
          "example": 0
        },
        "signatureOrProof": {
          "type": "string",
          "example": "string"
        }
      }
    },
...
    "ValidatorSetInstructionFilter": {
      "type": "object",
      "properties": {
        "remoteDestinationNetworkId": {
          "type": "number",
          "example": 0
        },
        "callbackURL": {
          "type": "string",
          "example": "string"
        }
      }
    },
    "ValidatorSetInstructionRequest": {
      "type": "object",
      "properties": {
        "operationId": {
          "type": "string",
          "example": "string"
        },
        "validators": {
          "type": "array",
          "example": [
            "string"
          ],
          "items": {
            "type": "string"
          }
        },
        "filters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidatorSetInstructionFilter"
          }
        }
      },
      "required": [
        "operationId",
        "validators"
      ]
    },
    "ValidatorSetInstructionResponse": {
      "type": "object",
      "properties": {
        "networkId": {
          "type": "number",
          "example": 0
        },
        "operationId": {
          "type": "string",
          "example": "string"
        },
        "state": {
          "type": "string",
          "example": "string"
        },
        "creationDate": {
          "type": "number",
          "example": 0
        },
        "lastUpdate": {
          "type": "number",
          "example": 0
        },
        "humanReadableTimestamp": {
          "type": "string",
          "example": "string"
        },
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EthereumProof"
          }
        }
      }
    }
  }
}